# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ECOnexão API
  description: API de domínio do aplicativo ECOnexão para turismo sustentável e rotas
    ecológicas.
  version: 1.0.0
paths:
...
  /api/v1/me:
    get:
      tags:
      - User - Profile & Preferences
      summary: Perfil do usuário atual
      description: Retorna os dados do perfil do usuário autenticado.
      operationId: get_my_profile_api_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileEnvelope'
      security:
      - HTTPBearer: []
    patch:
      tags:
      - User - Profile & Preferences
      summary: Atualizar perfil do usuário atual
      description: Atualiza campos do perfil do usuário autenticado.
      operationId: update_my_profile_api_v1_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/me/avatar:
    post:
      tags:
      - User - Profile & Preferences
      summary: Substituir avatar
      description: Recebe uma imagem multipart, sanitiza e gera derivados WebP no
        backend, publicando-os no Supabase Storage sem expor credenciais privilegiadas.
      operationId: replace_avatar_api_v1_me_avatar_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_replace_avatar_api_v1_me_avatar_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarUploadResponseEnvelope'
        '401':
          description: JWT ausente, inválido ou bloqueado por exclusão de conta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Imagem ausente, inválida, incompatível ou acima do limite.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
  /api/v1/me/account:
    delete:
      tags:
      - User - Profile & Preferences
      summary: Excluir a conta atual
      description: Remove avatares, dados de domínio e identidade Auth. É idempotente
        enquanto o JWT residual ainda for válido.
      operationId: delete_my_account_api_v1_me_account_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Saga de exclusão incompleta e segura para nova tentativa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
  /api/v1/me/preferences:
    get:
      tags:
      - User - Profile & Preferences
      summary: Preferências do usuário atual
      description: Retorna as preferências de acessibilidade, região e navegação do
        usuário.
      operationId: get_my_preferences_api_v1_me_preferences_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesEnvelope'
      security:
      - HTTPBearer: []
    patch:
      tags:
      - User - Profile & Preferences
      summary: Atualizar preferências do usuário atual
      description: Atualiza preferências de acessibilidade e região ativa do usuário.
      operationId: update_my_preferences_api_v1_me_preferences_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferencesUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/me/favorite-routes:
    get:
      tags:
      - User - Profile & Preferences
      summary: Rotas salvas pelo usuário atual
      description: Retorna a lista paginada de rotas favoritadas pelo usuário autenticado.
      operationId: get_my_favorite_routes_api_v1_me_favorite_routes_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Cursor opaco retornado em meta.next_cursor
          title: Cursor
        description: Cursor opaco retornado em meta.next_cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteListEnvelope'
        '401':
          description: Autenticação obrigatória.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Cursor inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/favorite-routes/{route_id}:
    put:
      tags:
      - User - Profile & Preferences
      summary: Salvar rota como favorita (Idempotente)
      description: Adiciona a rota aos favoritos do usuário de forma idempotente.
      operationId: add_favorite_route_api_v1_me_favorite_routes__route_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: route_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Route Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        '401':
          description: Autenticação obrigatória.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rota não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - User - Profile & Preferences
      summary: Remover rota dos favoritos (Idempotente)
      description: Remove a rota dos favoritos do usuário de forma idempotente.
      operationId: remove_favorite_route_api_v1_me_favorite_routes__route_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: route_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Route Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        '401':
          description: Autenticação obrigatória.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/me/favorite-actors:
    get:
      tags:
      - User - Profile & Preferences
      summary: Atores salvos pelo usuário atual
      description: Retorna a lista paginada de atores favoritados pelo usuário autenticado.
      operationId: get_my_favorite_actors_api_v1_me_favorite_actors_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Cursor opaco retornado em meta.next_cursor
          title: Cursor
        description: Cursor opaco retornado em meta.next_cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorListEnvelope'
        '401':
          description: Autenticação obrigatória.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Cursor inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/favorite-actors/{actor_id}:
    put:
      tags:
      - User - Profile & Preferences
      summary: Salvar ator como favorito (Idempotente)
      description: Adiciona o ator aos favoritos do usuário de forma idempotente.
      operationId: add_favorite_actor_api_v1_me_favorite_actors__actor_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: actor_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Actor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        '401':
          description: Autenticação obrigatória.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ator não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - User - Profile & Preferences
      summary: Remover ator dos favoritos (Idempotente)
      description: Remove o ator dos favoritos do usuário de forma idempotente.
      operationId: remove_favorite_actor_api_v1_me_favorite_actors__actor_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: actor_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Actor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        '401':
          description: Autenticação obrigatória.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/me/trips:
    get:
      tags:
      - User - Profile & Preferences
      summary: Histórico de viagens do usuário
      description: Retorna o histórico de viagens iniciadas e concluídas pelo usuário
        autenticado.
      operationId: get_my_trips_api_v1_me_trips_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripListEnvelope'
      security:
      - HTTPBearer: []
    post:
      tags:
      - User - Profile & Preferences
      summary: Iniciar nova viagem
      description: Cria um novo registro de viagem para o usuário autenticado na rota
        informada.
      operationId: create_trip_api_v1_me_trips_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/me/trips/{trip_id}/pause:
    post:
      tags:
      - User - Profile & Preferences
      summary: Pause Trip
      operationId: pause_trip_api_v1_me_trips__trip_id__pause_post
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripEnvelope'
      security:
      - HTTPBearer: []
  /api/v1/me/trips/{trip_id}/resume:
    post:
      tags:
      - User - Profile & Preferences
      summary: Resume Trip
      operationId: resume_trip_api_v1_me_trips__trip_id__resume_post
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripEnvelope'
      security:
      - HTTPBearer: []
  /api/v1/me/trips/{trip_id}/finish:
    post:
      tags:
      - User - Profile & Preferences
      summary: Finish Trip
      operationId: finish_trip_api_v1_me_trips__trip_id__finish_post
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripEnvelope'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    ActorListEnvelope:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ActorSummarySchema'
          type: array
          title: Data
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      type: object
      required:
      - data
      - meta
      title: ActorListEnvelope
...
    AvatarUploadResponseData:
      properties:
        media_asset_id:
          type: string
          format: uuid
          title: Media Asset Id
        url:
          type: string
          title: Url
        derivatives:
          additionalProperties:
            type: string
          type: object
          title: Derivatives
        alt_text:
          type: string
          title: Alt Text
      type: object
      required:
      - media_asset_id
      - url
      - derivatives
      - alt_text
      title: AvatarUploadResponseData
    AvatarUploadResponseEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/AvatarUploadResponseData'
      type: object
      required:
      - data
      title: AvatarUploadResponseEnvelope
...
    Body_replace_avatar_api_v1_me_avatar_post:
      properties:
        file:
          type: string
          format: binary
          title: File
          description: Imagem JPEG, PNG ou WebP; máximo 5 MiB
      type: object
      required:
      - file
      title: Body_replace_avatar_api_v1_me_avatar_post
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          example: NOT_FOUND
        message:
          type: string
          title: Message
          example: O recurso solicitado não foi encontrado.
        details:
          anyOf:
          - {}
          - type: 'null'
          title: Details
      type: object
      required:
      - code
      - message
      title: ErrorDetail
      description: Detailed information about an API error.
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        request_id:
          type: string
          title: Request Id
          example: req_123456789
      type: object
      required:
      - error
      - request_id
      title: ErrorResponse
      description: Standard error response envelope.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginationMeta:
      properties:
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - total
      - limit
      title: PaginationMeta
...
    RouteListEnvelope:
      properties:
        data:
          items:
            $ref: '#/components/schemas/RouteSummarySchema'
          type: array
          title: Data
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      type: object
      required:
      - data
      - meta
      title: RouteListEnvelope
...
    StandardSuccessData:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
      type: object
      title: StandardSuccessData
    StandardSuccessResponse:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          anyOf:
          - $ref: '#/components/schemas/StandardSuccessData'
          - type: 'null'
      type: object
      title: StandardSuccessResponse
...
    TripCreate:
      properties:
        route_id:
          type: string
          format: uuid
          title: Route Id
      type: object
      required:
      - route_id
      title: TripCreate
    TripEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/TripSchema'
      type: object
      required:
      - data
      title: TripEnvelope
    TripListEnvelope:
      properties:
        data:
          items:
            $ref: '#/components/schemas/TripSchema'
          type: array
          title: Data
      type: object
      required:
      - data
      title: TripListEnvelope
...
    UserPreferencesEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/UserPreferencesSchema'
      type: object
      required:
      - data
      title: UserPreferencesEnvelope
...
    UserPreferencesUpdate:
      properties:
        active_region_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Active Region Id
        screen_reader_mode:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Screen Reader Mode
        high_contrast:
          anyOf:
          - type: boolean
          - type: 'null'
          title: High Contrast
        text_scale:
          anyOf:
          - type: number
            maximum: 3.0
            minimum: 0.5
          - type: 'null'
          title: Text Scale
        locale:
          anyOf:
          - type: string
            maxLength: 10
          - type: 'null'
          title: Locale
      type: object
      title: UserPreferencesUpdate
    UserProfileEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/UserProfileSchema'
      type: object
      required:
      - data
      title: UserProfileEnvelope
    UserProfileSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        avatar_media_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Avatar Media Id
        avatar:
          anyOf:
          - $ref: '#/components/schemas/ResolvedMediaItemSchema'
          - type: 'null'
        status:
          type: string
          title: Status
          default: active
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      title: UserProfileSchema
    UserProfileUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
      type: object
      title: UserProfileUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer